// apps/api/src/modules/ai/ai.service.ts
export interface PatientHistoryData {
  patientId: string;
  patientName: string;
  diagnosticos: string[];
  tratamientos: string[];
  alergias: string[];
  notas: { content: string; createdAt: Date }[];
}